32 bit images

+++ SDHRManager.h
@@ -90,7 +90,7 @@ private:
 	struct ImageAsset {
 		void AssignByFilename(const char* filename);	// currently unused
 		void AssignByMemory(const uint8_t* buffer, uint64_t size);
-		void ExtractTile(uint16_t* tile_p,
+		void ExtractTile(uint32_t* tile_p,
 			uint16_t tile_xdim, uint16_t tile_ydim, uint64_t xsource, uint64_t ysource);
 		// image assets are full 32-bit bitmap files, uploaded from PNG
 		uint64_t image_xcount = 0;
@@ -107,7 +107,7 @@ private:
 		uint64_t xdim;
 		uint64_t ydim;
 		uint64_t num_entries;
-		uint16_t* tile_data = NULL;  // tiledata is 16-bit RGB
+		uint32_t* tile_data = NULL;  // tiledata is 32-bit RGB
 		TilesetRecord()
 			: xdim(0)
 			, ydim(0)
